refactor: extract reusable component from AdvancementLevel
