+++ frontend/src/components/common/ExpandableSection/ExpandableSection.interface.ts
@@ -1,0 +1,7 @@
+export interface ExpandableSectionProps {
+  title: string;
+  description?: string;
+  showVerticalLine?: boolean;
+  open: boolean;
+  onClick: () => void;
+}
