+++ frontend/src/components/common/ExpandableSection/index.ts
@@ -1,0 +1,1 @@
+export { ExpandableSection } from './ExpandableSection';
