+++ frontend/src/components/modules/AdvancementLevel/AdvancementLevel.hooks.ts
