--- frontend/src/components/modules/AdvancementLevel/AdvancementLevel.interface.ts
@@ -1,3 +1,0 @@
-export interface AdvancementLevelProps {
-  level?: number;
-}
