+++ frontend/src/components/modules/AdvancementLevel/AdvancementLevel.tsx
@@ -1,17 +1,61 @@
-import { AdvancementLevelProps } from './AdvancementLevel.interface';
-import { MAX_LEVEL, MIN_LEVEL } from './constants';
+'use client';
+import { AccordionCard } from '@app/components/common/AccordionCard';
+import { AccordionList } from '@app/components/common/AccordionList';
+import { Modal } from '@app/components/common/Modal';
+import { AdvancementLevelProps } from './AdvancemetLevel.interface';
+import { useAdvancementLevel } from './AdvancementLevel.hooks';
+import { Markdown } from '@app/components/common/Markdown';
+import { Button } from '@app/components/common/Button';
+import { ExpandableSection } from '@app/components/common/ExpandableSection';
 
-export const AdvancementLevel = ({ level }: AdvancementLevelProps) => {
-  const currentLevel = Math.min(Math.max(level ?? MIN_LEVEL, MIN_LEVEL), MAX_LEVEL);
+export const AdvancementLevel: React.FC<AdvancementLevelProps> = ({ showVerticalLine, data, open, onClick }) => {
+  const { hideModal, openModal, modalOpen } = useAdvancementLevel();
+
+  const { advancementLevel, description, projects, categories } = data;
+  const shouldBeExpandedByDefault = Object.keys(data.categories).length === 1;
 
   return (
-    <div className="flex flex-row gap-[7px]">
-      {[...Array(currentLevel)].map((_, idx) => (
-        <div key={`green-${idx}`} className="h-3 w-3 rounded-full bg-green-600"></div>
-      ))}
-      {[...Array(MAX_LEVEL - currentLevel)].map((_, idx) => (
-        <div key={`grey-${idx}`} className="h-3 w-3 rounded-full bg-navy-300"></div>
-      ))}
-    </div>
+    <ExpandableSection
+      title={`Advancement level ${advancementLevel}`}
+      open={open}
+      onClick={onClick}
+      description={description}
+      showVerticalLine={showVerticalLine}
+    >
+      <>
+        {projects.length > 0 && (
+          <Button variant="link" onClick={openModal} className="w-fit text-sm">
+            An example way to pass level
+          </Button>
+        )}
+        {Object.entries(categories).map(([category, skills]) => (
+          <AccordionCard
+            key={category}
+            className="w-full"
+            title={category}
+            expandedByDefault={shouldBeExpandedByDefault}
+            small
+          >
+            <AccordionList
+              items={skills.map(({ name, description }) => ({
+                key: name,
+                title: name,
+                children: description ? <p>{description}</p> : undefined,
+              }))}
+            />
+          </AccordionCard>
+        ))}
+        <Modal open={modalOpen} onClose={hideModal} title="An example way to pass level">
+          {projects.map(({ title, overview }) => (
+            <div key={title} className="overflow-hidden text-base text-navy-600">
+              <p>{title}</p>
+              <article className="prose mt-5">
+                <Markdown text={overview} />
+              </article>
+            </div>
+          ))}
+        </Modal>
+      </>
+    </ExpandableSection>
   );
 };
