+++ frontend/src/components/modules/LevelDots/LevelDots.interface.ts
@@ -1,0 +1,3 @@
+export interface LevelDotsProps {
+  level?: number;
+}
