+++ frontend/src/components/modules/LevelDots/constants.ts
