+++ frontend/src/components/modules/LevelDots/index.ts
@@ -1,0 +1,1 @@
+export { LevelDots } from './LevelDots';
