--- frontend/src/components/pages/BucketDetails/modules/AdvancementLevel/index.ts
@@ -1,1 +1,0 @@
-export { AdvancementLevel } from './AdvancementLevel';
